requirements.txt
streamlit
requests
fastapi
uvicorn[standard]
transformers
Pillow
python-multipart
torch
keras==2.12.0
numpy
tensorflow-data-validation==1.14.0
tensorflow-metadata==1.14.0
ipython==8.12.1
numpy==1.23.5
pandas==1.5.3
scikit-learn==1.3.0
gunicorn==20.1.0
flask==2.3.2
kfp==2.0.1